refactor: ErrorResponse Enum 이동
- exception 패키지로 이동
